moved Delimiters to common
