--- ayupov.dmitry/T3/delimiter.cpp
@@ -1,18 +1,0 @@
-#include "delimiter.hpp"
-#include <istream>
-
-std::istream& ayupov::operator>>(std::istream& in, DelimiterChar&& exp)
-{
-  std::istream::sentry guard(in);
-  if (!guard)
-  {
-    return in;
-  }
-  char c = 0;
-  in >> c;
-  if (c != exp.expected)
-  {
-    in.setstate(std::ios::failbit);
-  }
-  return in;
-}
