--- ayupov.dmitry/T3/delimiter.hpp
@@ -1,14 +1,0 @@
-#ifndef DELIMITER_HPP
-#define DELIMITER_HPP
-
-#include <istream>
-
-namespace ayupov
-{
-  struct DelimiterChar
-  {
-    char expected;
-  };
-  std::istream& operator>>(std::istream& in, DelimiterChar&& exp);
-}
-#endif
