+++ ayupov.dmitry/common/Delimiter.cpp
