+++ ayupov.dmitry/common/Delimiter.hpp
